code/requirements.txt
Flask-Assets
Flask
Flask-SQLAlchemy
sqlalchemy
pyScss
cssmin
Flask-HTTPAuth
gunicorn
httplib2
mechanicalsoup
PyDrive
#PyOpenSSL
pyasn1==0.4.2
python-telegram-bot
telegram
flask_apscheduler
apscheduler
requests
urllib3
# Gapps requirements
google-auth
google-auth-oauthlib
google-auth-httplib2
git+https://github.com/matteogaito/sprea-utils
